+++ dist-fs/comms/i2c.c
@@ -5,11 +5,11 @@
 static int i2c_init(comm_context_t *ctx);
 static int i2c_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int
-i2c_read(comm_context_t *ctx, uint8_t *rx, uint16_t rx_sz, uint16_t timeout_ms);
+i2c_read(comm_context_t *ctx, uint8_t *rx, uint32_t rx_sz, uint16_t timeout_ms);
 static int i2c_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int i2c_write(comm_context_t *ctx,
                      uint8_t *tx,
-                     uint16_t tx_size,
+                     uint32_t tx_size,
                      uint16_t timeout_ms);
 static int i2c_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
@@ -40,7 +40,7 @@ static int i2c_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 
 static int i2c_read(comm_context_t *ctx,
                     uint8_t *rx,
-                    uint16_t rx_sz,
+                    uint32_t rx_sz,
                     uint16_t timeout_ms) {
   (void)ctx;
   (void)rx;
@@ -63,7 +63,7 @@ static int i2c_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
 
 static int i2c_write(comm_context_t *ctx,
                      uint8_t *tx,
-                     uint16_t tx_size,
+                     uint32_t tx_size,
                      uint16_t timeout_ms) {
   (void)ctx;
   (void)tx;
