+++ dist-fs/comms/network.c
@@ -6,13 +6,13 @@ static int network_init(comm_context_t *ctx);
 static int network_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int network_read(comm_context_t *ctx,
                         uint8_t *rx,
-                        uint16_t rx_sz,
+                        uint32_t rx_sz,
                         uint16_t timeout_ms);
 static int
 network_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int network_write(comm_context_t *ctx,
                          uint8_t *tx,
-                         uint16_t tx_size,
+                         uint32_t tx_size,
                          uint16_t timeout_ms);
 static int network_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
@@ -42,7 +42,7 @@ static int network_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 
 static int network_read(comm_context_t *ctx,
                         uint8_t *rx,
-                        uint16_t rx_sz,
+                        uint32_t rx_sz,
                         uint16_t timeout_ms) {
   (void)ctx;
   (void)rx;
@@ -66,7 +66,7 @@ network_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
 
 static int network_write(comm_context_t *ctx,
                          uint8_t *tx,
-                         uint16_t tx_size,
+                         uint32_t tx_size,
                          uint16_t timeout_ms) {
   (void)ctx;
   (void)tx;
