+++ dist-fs/comms/spi.c
@@ -5,11 +5,11 @@
 static int spi_init(comm_context_t *ctx);
 static int spi_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int
-spi_read(comm_context_t *ctx, uint8_t *rx, uint16_t rx_sz, uint16_t timeout_ms);
+spi_read(comm_context_t *ctx, uint8_t *rx, uint32_t rx_sz, uint16_t timeout_ms);
 static int spi_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int spi_write(comm_context_t *ctx,
                      uint8_t *tx,
-                     uint16_t tx_size,
+                     uint32_t tx_size,
                      uint16_t timeout_ms);
 static int spi_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
@@ -39,7 +39,7 @@ static int spi_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 
 static int spi_read(comm_context_t *ctx,
                     uint8_t *rx,
-                    uint16_t rx_sz,
+                    uint32_t rx_sz,
                     uint16_t timeout_ms) {
   (void)ctx;
   (void)rx;
@@ -62,7 +62,7 @@ static int spi_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
 
 static int spi_write(comm_context_t *ctx,
                      uint8_t *tx,
-                     uint16_t tx_size,
+                     uint32_t tx_size,
                      uint16_t timeout_ms) {
   (void)ctx;
   (void)tx;
