+++ dist-fs/comms/uart.c
@@ -16,12 +16,12 @@ static int uart_init(comm_context_t *ctx);
 static int uart_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int uart_read(comm_context_t *ctx,
                      uint8_t *rx,
-                     uint16_t rx_sz,
+                     uint32_t rx_sz,
                      uint16_t timeout_ms);
 static int uart_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int uart_write(comm_context_t *ctx,
                       uint8_t *tx,
-                      uint16_t tx_size,
+                      uint32_t tx_size,
                       uint16_t timeout_ms);
 static int uart_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
@@ -134,7 +134,7 @@ static int uart_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 
 static int uart_read(comm_context_t *ctx,
                      uint8_t *rx,
-                     uint16_t rx_sz,
+                     uint32_t rx_sz,
                      uint16_t timeout_ms) {
   if (!ctx || !ctx->driver || !rx)
     return -EINVAL;
@@ -185,7 +185,7 @@ uart_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
 
 static int uart_write(comm_context_t *ctx,
                       uint8_t *tx,
-                      uint16_t tx_size,
+                      uint32_t tx_size,
                       uint16_t timeout_ms) {
   if (!ctx || !ctx->driver || !tx)
     return -EINVAL;
